backend/requirements.txt
alembic==1.11.3
amqp==5.2.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1
bcrypt==4.2.0
billiard==4.2.1
blinker==1.6.2
celery==5.4.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.36.0
exceptiongroup==1.2.0
Flask==2.2.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.4.2
Mako==1.1.0
MarkupSafe==2.1.2
packaging==23.2
Pillow==9.0.1
pluggy==1.3.0
prompt-toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.7.0
pytest==7.4.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.8
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.50
tinify==1.6.0
tomli==2.0.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
zipp==3.15.0
